fixed minor formatting issues
