+++ src/App.css
@@ -2,6 +2,10 @@
   padding-top: 3em;
 }
 
+.changeTermsForm {
+  
+}
+
 .input-field, .add-form-button {
   padding-bottom: 1em;
 }
